+++ src/script/util/LocalizerUtil.test.ts
